backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
pydantic==2.7.1
httpx==0.27.0
python-dateutil==2.9.0.post0
alembic==1.13.1
